[Wearable] Add EventHandler to RotarySelector
"item,selected": When the user selected an item.
"item,clicked": When selecting again the alredy selected item or selecting a selector.
"editing,entered": When the user entered to the editing mode.
"editing,exited": When the user exited from the editing mode.

Need to add event
"item,deleted": When the user deleted an item.
"item,reordered": When the user reordered an item.

Signed-off-by: huiyu.eun <huiyu.eun@samsung.com>